rename: signal_handler.h → signal.h、シグナル関数をファイル分割
