+++ Makefile
@@ -133,7 +133,8 @@ SRCS_MAND	:=	src/main.c \
 				src/expand/expand_wildcard/resolve_wildcard.c \
 				src/expand/expand_wildcard/expand_wildcard.c \
 				src/expand/expand_remove_quotes/expand_remove_quotes.c \
-				src/signal/setup_signal.c \
+				src/signal/setup_signal_handlers.c \
+				src/signal/set_signals.c \
 
 SRCS_BONUS	:=
