+++ includes/signal.h
@@ -1,0 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: surayama <surayama@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/04/06 00:00:00 by surayama          #+#    #+#             */
+/*   Updated: 2026/04/06 00:00:00 by surayama         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef SIGNAL_H
+# define SIGNAL_H
+
+# include <signal.h>
+
+extern volatile sig_atomic_t	g_signal;
+
+void	setup_signal_handlers(void);
+void	set_signal_ignore(void);
+void	set_signal_default(void);
+void	set_signal_interactive(void);
+
+#endif
