+++ src/prompt/prompt.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 
 #include "prompt.h"
-#include "signal_handler.h"
+#include "signal.h"
 
 void	prompt(void (*handler)(char *input, t_shell_table *shell_table),
 		t_shell_table *shell_table)
